requirements.txt
google-genai
python-dotenv
requests
beautifulsoup4
faiss-cpu
numpy
streamlit
